forms: fix text area default value
We can't use `value` on TextAreaTag because that will just add it as an attribute, but the content of the textarea needs to be in the content of the tag, not the value attribute.

After the tag has been rendered, the actual element object supports setting/getting with the value property.