Part: add bound child parts
bound parts allow for implicitly defining the state of a child part based on its parent's state. this is a two way connection, so when the state of the parent changes, the state of the child will also update, and when the state of the child updates, it will pass those updates back up to its parent.