Nav: add tryVisit
this is useful for the case where the captured portion of the app is relatively small and you want to navigate using tuff if possible, but want to fall back to app defined navigation if the given path isn't captured.

The motivating usecase for this is clypboard which uses Turbolinks normally but we want to bypass turbolinks navigation for tuff pages and fall back to the app-defined turbolinks navigation for non-captured paths.